% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar}
\alias{calendar}
\alias{calendar_get_events}
\alias{calendar_unavailable_users}
\title{Connect to Calendar in Outlook 365}
\usage{
calendar_get_events(
  start_date = Sys.Date(),
  end_date = start_date,
  account = connect_outlook()
)

calendar_unavailable_users(
  start_date = Sys.Date(),
  end_date = start_date,
  account = connect_outlook()
)
}
\arguments{
\item{start_date, end_date}{date range}

\item{account}{the Microsoft 365 Outlook account object, e.g. as returned by \code{\link[=connect_outlook]{connect_outlook()}}}
}
\description{
After connecting using \code{\link[=connect_outlook]{connect_outlook()}}, the \verb{get_*()} functions retrieve properties from this connection.
}
\details{
\code{\link[=calendar_get_events]{calendar_get_events()}} retrieves a \link{data.frame} with all appointments from the shared calendar set in \code{account}.

\code{\link[=calendar_unavailable_users]{calendar_unavailable_users()}} returns a vector of appointment subjects that are shown as "free" within \code{start_date} and \code{end_date}.
}
