% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certepackage.R
\docType{package}
\name{certemail-package}
\alias{certemail}
\alias{certemail-package}
\title{certemail: A Certe R Package for Automated Mailing}
\description{
A Certe R Package for working remotely with emails, using Microsoft 365.
  This package is part of the 'certedata' universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://certe-medical-epidemiology.github.io/certemail}
  \item \url{https://github.com/certe-medical-epidemiology/certemail}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.berends@certe.nl}

Authors:
\itemize{
  \item Erwin E. A. Hassing \email{e.hassing@certe.nl}
}

Other contributors:
\itemize{
  \item Certe Medical Diagnostics & Advice Foundation [copyright holder, funder]
}

}
\keyword{internal}
