% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_properties.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{connect_outlook}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{certeprojects}{\code{\link[certeprojects:connect]{connect_outlook}}}
}}

