% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_properties.R
\name{account_properties}
\alias{account_properties}
\alias{connect_outlook365}
\alias{get_name}
\alias{get_name_and_job_title}
\alias{get_name_and_mail_address}
\alias{get_mail_address}
\alias{get_inbox_name}
\title{Connect to Outlook Business 365}
\usage{
connect_outlook365(tenant = read_secret("mail.tenant"), error_on_fail = FALSE)

get_name()

get_name_and_job_title()

get_name_and_mail_address()

get_mail_address()

get_inbox_name()
}
\arguments{
\item{tenant}{the tenant to use for \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}}

\item{error_on_fail}{a \link{logical} to indicate whether an error must be thrown if no connection can be made}
}
\description{
This function creates a connection to Outlook Business 365 and saves this connection to the \code{certemail} package environment. The \verb{get_*()} functions retrieve properties from this connection.
}
