% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_properties.R
\name{account_properties}
\alias{account_properties}
\alias{get_name}
\alias{get_department}
\alias{get_address}
\alias{get_phone_numbers}
\alias{get_mail_address}
\alias{get_inbox_name}
\alias{get_drafts_name}
\alias{get_certe_name}
\alias{get_job_title}
\alias{get_certe_name_and_job_title}
\alias{get_certe_location}
\alias{get_certe_signature}
\title{Work with Outlook 365}
\usage{
get_name(account = connect_outlook())

get_department(account = connect_outlook())

get_address(account = connect_outlook())

get_phone_numbers(account = connect_outlook())

get_mail_address(account = connect_outlook())

get_inbox_name(account = connect_outlook())

get_drafts_name(account = connect_outlook())

get_certe_name(user = Sys.info()["user"], account = connect_outlook())

get_job_title(user = Sys.info()["user"], account = connect_outlook())

get_certe_name_and_job_title(
  user = Sys.info()["user"],
  account = connect_outlook()
)

get_certe_location(account = connect_outlook())

get_certe_signature(account = connect_outlook(), plain = FALSE)
}
\arguments{
\item{account}{the Microsoft 365 Outlook account object, e.g. as returned by \code{\link[=connect_outlook]{connect_outlook()}}}

\item{user}{Certe user ID, to look up the \link[certetoolbox:read_secret]{secret} \verb{user.\{user\}.fullname}}

\item{plain}{a \link{logical} to indicate whether textual formatting should not be applied}
}
\description{
After connecting using \code{\link[=connect_outlook]{connect_outlook()}}, the \verb{get_*()} functions retrieve properties from this connection.
}
\details{
The \code{\link[=get_certe_name]{get_certe_name()}} and \code{\link[=get_certe_name_and_job_title]{get_certe_name_and_job_title()}} functions first look for the \link[certetoolbox:read_secret]{secret} \verb{user.\{user\}.fullname} and falls back to \code{\link[=get_name]{get_name()}} if it is not available.
}
