% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_properties.R
\name{account_properties}
\alias{account_properties}
\alias{connect_outlook365}
\alias{get_name}
\alias{get_name_and_job_title}
\alias{get_name_and_mail_address}
\alias{get_mail_address}
\alias{get_department}
\alias{get_inbox_name}
\alias{get_drafts_name}
\title{Connect to Outlook Business 365}
\usage{
connect_outlook365(tenant = read_secret("mail.tenant"), error_on_fail = FALSE)

get_name(account = connect_outlook365())

get_name_and_job_title(account = connect_outlook365())

get_name_and_mail_address(account = connect_outlook365())

get_mail_address(account = connect_outlook365())

get_department(account = connect_outlook365())

get_inbox_name(account = connect_outlook365())

get_drafts_name(account = connect_outlook365())
}
\arguments{
\item{tenant}{the tenant to use for \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}}

\item{error_on_fail}{a \link{logical} to indicate whether an error must be thrown if no connection can be made}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=connect_outlook365]{connect_outlook365()}} or \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}.}
}
\description{
This function creates a connection to Outlook Business 365 and saves this connection to the \code{certemail} package environment. The \verb{get_*()} functions retrieve properties from this connection.
}
