% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_properties.R
\name{account_properties}
\alias{account_properties}
\alias{connect_outlook365}
\alias{get_name}
\alias{get_department}
\alias{get_address}
\alias{get_phone_numbers}
\alias{get_mail_address}
\alias{get_inbox_name}
\alias{get_drafts_name}
\alias{get_certe_name}
\alias{get_job_title}
\alias{get_certe_name_and_job_title}
\alias{get_certe_location}
\alias{get_certe_signature}
\title{Connect to Outlook Business 365}
\usage{
connect_outlook365(
  shared_mbox_email = read_secret("mail.auto_from"),
  tenant = read_secret("mail.tenant"),
  app_id = read_secret("mail.app_id"),
  auth_type = read_secret("mail.auth_type"),
  error_on_fail = FALSE
)

get_name(account = connect_outlook365())

get_department(account = connect_outlook365())

get_address(account = connect_outlook365())

get_phone_numbers(account = connect_outlook365())

get_mail_address(account = connect_outlook365())

get_inbox_name(account = connect_outlook365())

get_drafts_name(account = connect_outlook365())

get_certe_name(user = Sys.info()["user"], account = connect_outlook365())

get_job_title(user = Sys.info()["user"], account = connect_outlook365())

get_certe_name_and_job_title(
  user = Sys.info()["user"],
  account = connect_outlook365()
)

get_certe_location(account = connect_outlook365())

get_certe_signature(account = connect_outlook365(), plain = FALSE)
}
\arguments{
\item{tenant}{the tenant to use for \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}}

\item{app_id}{the Azure app id to use for \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}}

\item{error_on_fail}{a \link{logical} to indicate whether an error must be thrown if no connection can be made}

\item{account}{a Microsoft 365 account to use for looking up properties. This has to be an object as returned by \code{\link[=connect_outlook365]{connect_outlook365()}} or \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}.}

\item{user}{Certe user ID, to look up the \link[certetoolbox:read_secret]{secret} \verb{user.\{user\}.fullname}}

\item{plain}{a \link{logical} to indicate whether textual formatting should not be applied}
}
\description{
These functions create a connection to Outlook Business 365 and saves the connection to the \code{certemail} package environment. The \verb{get_*()} functions retrieve properties from this connection.
}
\details{
The \code{\link[=get_certe_name]{get_certe_name()}} and \code{\link[=get_certe_name_and_job_title]{get_certe_name_and_job_title()}} functions first look for the \link[certetoolbox:read_secret]{secret} \verb{user.\{user\}.fullname} and falls back to \code{\link[=get_name]{get_name()}} if it is not available.
}
