% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mail_attachment.R
\name{download_mail_attachment}
\alias{download_mail_attachment}
\title{Download Email Attachments Using Microsoft 365}
\usage{
download_mail_attachment(
  path = getwd(),
  filename = "mail_{name}_{date}_{time}",
  search = "hasattachment:yes",
  search_subject = NULL,
  search_from = NULL,
  search_when = NULL,
  folder = get_inbox_name(),
  n = 5,
  sort = "received desc",
  overwrite = TRUE
)
}
\arguments{
\item{path}{location to save attachment(s) in}

\item{filename}{new filename for the attachments, use \code{NULL} to not alter the filename. The text \code{"{date}"} will be replaced with the date of the mail in format YYYYMMDD, \code{"{time}"} will be replaced with the time of the mail in format HHMMSS, and \code{"{name}"} will be replaced with the sender name. Invalid filename characters will be replaced with an underscore.}

\item{search}{an ODATA filter, ignores \code{sort} and defaults to search only mails with attachments}

\item{search_subject}{a \link{character}, equal to \code{search = "subject:(search_subject)"}, case-insensitive}

\item{search_from}{a \link{character}, equal to \code{search = "from:(search_from)"}, case-insensitive}

\item{search_when}{a \link{Date} vector of size 1 or 2, equal to \code{search = "received:date1..date2"}, see \emph{Examples}}

\item{folder}{email folder to search in, defaults to Inbox name of the current user by calling \code{\link[=get_inbox_name]{get_inbox_name()}}}

\item{n}{maximum number of emails to choose from}

\item{sort}{initial sorting}

\item{overwrite}{logical to indicate whether existing local files should be overwritten}
}
\description{
This uses the \code{Microsoft365R} package to download email attachments via Microsoft 365. Connection will be made internally using \code{\link[Microsoft365R:client]{get_business_outlook()}} using the Certe tenant ID.
}
\details{
If both \code{search_subject} and \code{search_from} are provided, they will be matched as 'AND'. \code{search_from} can contain any sender name or email address. If \code{search_when} has a length over 2, the first and last value will be taken.

See this page for all filtering options using the \code{search} argument: \url{https://support.microsoft.com/en-gb/office/how-to-search-in-outlook-d824d1e9-a255-4c8a-8553-276fb895a8da}.
}
\examples{
\dontrun{
download_mail_attachment(search_when = "2021-10-28")

if (require("certetoolbox")) {
  download_mail_attachment(search_from = "glims", search_when = last_week())

  download_mail_attachment(search_from = "drenthe", search_when = yesterday(), n = 1)
}
}
}
\seealso{
\code{\link[=mail]{mail()}}
}
