% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mail_attachment.R
\name{download_mail_attachment}
\alias{download_mail_attachment}
\title{Download Email Attachments Using Microsoft 365}
\usage{
download_mail_attachment(
  path = getwd(),
  filename = "mail_{name}_{datetime}",
  search = "hasattachment:yes",
  search_subject = NULL,
  search_from = NULL,
  search_when = NULL,
  search_attachment = NULL,
  folder = get_inbox_name(account = account),
  n = 5,
  sort = "received desc",
  overwrite = TRUE,
  skip_inline = TRUE,
  account = connect_outlook(),
  interactive_mode = interactive()
)
}
\arguments{
\item{path}{location to save attachment(s) to}

\item{filename}{new filename for the attachments, use \code{NULL} to not alter the filename. For multiple files, the files will be appended with a number. The following texts can be used for replacement (invalid filename characters will be replaced with an underscore):
\itemize{
\item \code{"{date}"}, the received date of the email in format "yyyymmdd"
\item \code{"{time}"}, the received time of the email in format "HHMMSS"
\item \code{"{datetime}"} and \code{"{date_time}"}, the received date and time of the email in format "yyyymmdd_HHMMSS"
\item \code{"{name}"}, the name of the sender of the email
\item \code{"{address}"}, the email address of the sender of the email
\item \code{"{original}"}, the original filename of the attachment
}}

\item{search}{an ODATA filter, ignores \code{sort} and defaults to search only mails with attachments}

\item{search_subject}{a \link{character}, equal to \code{search = "subject:(search_subject)"}, case-insensitive}

\item{search_from}{a \link{character}, equal to \code{search = "from:(search_from)"}, case-insensitive}

\item{search_when}{a \link{Date} vector of size 1 or 2, equal to \code{search = "received:date1..date2"}, see \emph{Examples}}

\item{search_attachment}{a \link{character} to use a regular expression for attachment file names}

\item{folder}{email folder name to search in, defaults to Inbox of the current user by calling \code{\link[=get_inbox_name]{get_inbox_name()}}}

\item{n}{maximum number of emails to search}

\item{sort}{initial sorting}

\item{overwrite}{a \link{logical} to indicate whether existing local files should be overwritten}

\item{skip_inline}{a \link{logical} to indicate whether inline attachments such as meta images must be skipped}

\item{account}{a Microsoft 365 account to use for searching the mails. This has to be an object as returned by \code{\link[=connect_outlook]{connect_outlook()}} or \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}.}

\item{interactive_mode}{a \link{logical} to indicate interactive mode. In non-interactive mode, all attachments within the search criteria will be downloaded.}
}
\description{
This uses the \code{Microsoft365R} package to download email attachments via Microsoft 365. Connection will be made using \code{\link[=connect_outlook]{connect_outlook()}}.
}
\details{
\verb{search_*} arguments will be matched as 'AND'.

\code{search_from} can contain any sender name or email address. If \code{search_when} has a length over 2, the first and last value will be taken.

Refer to \href{https://support.microsoft.com/en-gb/office/how-to-search-in-outlook-d824d1e9-a255-4c8a-8553-276fb895a8da}{this Microsoft Support page} for all filtering options using the \code{search} argument.
}
\examples{
\dontrun{
download_mail_attachment(search_when = "2021-10-28")

if (require("certetoolbox")) {
  download_mail_attachment(search_from = "groningen",
                           search_when = last_week(),
                           filename = "groningen_{date_time}")

  download_mail_attachment(search_from = "drenthe",
                           search_when = yesterday(),
                           n = 1)
}
}
}
\seealso{
\code{\link[=mail]{mail()}}
}
