% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{mail}
\alias{mail}
\alias{mail_plain}
\alias{mail_dataset}
\alias{mail_image}
\alias{mail_on_error}
\alias{download_mail_attachment}
\title{Send emails using Microsoft 365}
\usage{
mail(
  body,
  subject,
  to,
  cc = NULL,
  bcc = read_secret("mail.auto_bcc"),
  reply_to = NULL,
  attachment = NULL,
  header = FALSE,
  footer = FALSE,
  background = colourpicker("certeblauw3"),
  send = TRUE,
  md_body = TRUE,
  signature = TRUE,
  signature_name = get_name_and_job_title(),
  signature_address = get_mail_address(),
  automated_notice = !interactive(),
  outlook = FALSE,
  save_location = read_secret("mail.export_path"),
  expect = NULL,
  ...
)

mail_plain(
  body,
  subject,
  to,
  cc = NULL,
  bcc = read_secret("mail.auto_bcc"),
  reply_to = read_secret("mail.auto_reply_to"),
  attachment = NULL,
  send = TRUE,
  ...
)

mail_dataset(df, font.size = 11, ..., plain = FALSE)

mail_image(image_path, width = NULL, ...)

mail_on_error(expr, to = read_secret("mail.error_to"), ...)

download_mail_attachment(
  path = getwd(),
  search = NULL,
  folder = "Postvak IN",
  n = 50,
  sort = "received desc",
  overwrite = TRUE
)
}
\arguments{
\item{body}{Tekst van de e-mail.}

\item{subject}{Onderwerp van de e-mail. Spaties aan begin en einde worden weggesneden.}

\item{to}{Geaddresseerde (veld 'Aan') van de e-mail. Kan ook vector zijn. Standaard het e-mailadres zoals gedefinieerd in de omgevingsvariabele van de ingelogde gebruiker.}

\item{cc}{Geaddresseerde (veld 'CC') van de e-mail. Kan ook vector zijn.}

\item{bcc}{Geaddresseerde (veld 'BCC') van de e-mail. Kan ook vector zijn. Adressen die voorkomen in \code{to} of \code{cc} worden eruit gehaald. Als in \code{to} of \code{cc} minimaal 1 data-onderzoeker staat, wordt dit standaard niet naar de data-onderzoekers gestuurd als BCC.}

\item{reply_to}{Geaddresseerde (veld 'Van') van de e-mail. In Outlook moeten de rechten vastgelegd zijn om vanaf dit account te versturen.}

\item{attachment}{Bijlagen: vector van locaties. Als de locatie niet bestaat, wordt er een fout gegeven en wordt de mail niet verstuurd.}

\item{header, footer}{Extra info voor boven en onder de body. Zie \link{blocks} voor het maken van blokken die hier geplaatst kunnen worden. Bij \code{FALSE} worden deze blokken verborgen.}

\item{background}{Achtergrond van de mail (standaard: \code{\link{colourpicker}("certeblauw3")}). Gebruik \code{""}, \code{NULL} of \code{FALSE} om geen achtergrond te gebruiken.}

\item{send}{Direct versturen van de e-mail. Met \code{send = FALSE} wordt de e-mail op het scherm weergegeven. Standaard \code{TRUE}.}

\item{md_body}{Tekst van de e-mail als Markdown opmaken (standaard: \code{TRUE}).}

\item{signature}{Handtekening aan e-mail toevoegen (standaard: \code{TRUE}).}

\item{signature_name}{Naam die onder 'Met vriendelijke groet' komt te staan.}

\item{signature_address}{E-mailaderes die onder 'Met vriendelijke groet' komt te staan.}

\item{automated_notice}{Onderaan de mail deze tekst weergeven: "\emph{Deze mail is geautomatiseerd verstuurd.}" (standaard: \code{TRUE}).}

\item{outlook}{Mail opstellen met Outlook (standaard: \code{FALSE}). Anders wordt de mail rechtstreeks via de SMTP-server verzonden.}

\item{save_location}{De maplocatie waar de mail opgeslagen wordt nadat deze verzonden is. Deze wordt opgeslagen als .rds-bestand (met evt. bijlagen), die in RStudio geopend kan worden en geprint kan worden in de Console voor alle maildetails. De mail kan dan zelfs opnieuw verstuurd worden. Gebruik \code{""}, \code{NULL} of \code{FALSE} om de mail niet op te slaan.}

\item{expect}{Standaard is leeg. Een expressie die \code{TRUE} moet zijn alvorens te mailen. Als dit \code{FALSE} is om een fout oplevert, wordt \code{mail_on_error()} gedraaid met details over de fout.}

\item{...}{Voor \code{mail_on_error}: parameters die doorgegeven worden aan \code{mail()}. \cr Voor \code{mail_dataset()}: parameters die doorgegeven worden aan \code{\link{tbl_flextable}()}.}

\item{df}{Een dataset}

\item{font.size}{Standaard is 11. Lettergrootte die doorgegeven wordt aan \code{\link{tbl_flextable}()}.}

\item{image_path}{path of image}

\item{width}{required width of image, must be in CSS style such as "200px"}

\item{expr}{Expressie die gedraaid moet worden}

\item{path}{save location for attachment(s)}

\item{search}{an ODATA filter, ignores \code{sort}}

\item{folder}{email folder to search in}

\item{n}{maximum number of email to list}

\item{sort}{initial sorting}

\item{overwrite}{logical to indicate whether existing local files should be overwritten}
}
\description{
Send emails using Microsoft 365
}
\details{
De functie \code{mail_on_error()} is speciaal voor gebruik in geautomatiseerde scripts. Het evalueert de input en verstuurt een mail met de expressie wanneer deze een fout retourneert.

De functies \code{mail_dataset()} en \code{mail_image()} zijn om resp. een tabel met \code{\link{tbl_flextable}()} en een afbeelding in de mail te plaatsen.
}
