% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{mail}
\alias{mail}
\alias{mail_plain}
\alias{mail_image}
\alias{mail_on_error}
\alias{mail_is_sent}
\title{Send Emails Using Microsoft 365}
\usage{
mail(
  body,
  subject = "",
  to = NULL,
  cc = read_secret("mail.auto_cc"),
  bcc = read_secret("mail.auto_bcc"),
  reply_to = NULL,
  attachment = NULL,
  header = FALSE,
  footer = FALSE,
  background = certestyle::colourpicker("certeblauw6"),
  send = TRUE,
  markdown = TRUE,
  signature = get_certe_signature(account = account),
  automated_notice = !interactive(),
  save_location = read_secret("mail.export_path"),
  sent_subfolder = read_secret("mail.sent_subfolder"),
  expect = NULL,
  account = connect_outlook(),
  account_teams = connect_teams(),
  identifier = NULL,
  ...
)

mail_plain(
  body,
  subject = "",
  to = NULL,
  cc = read_secret("mail.auto_cc"),
  bcc = read_secret("mail.auto_bcc"),
  reply_to = read_secret("mail.auto_reply_to"),
  attachment = NULL,
  send = TRUE,
  ...
)

mail_image(image_path, width = NULL, ...)

mail_on_error(expr, to = read_secret("mail.error_to"), ...)

mail_is_sent(project_number, date = Sys.Date(), account = connect_outlook())
}
\arguments{
\item{body}{body of email, allows markdown if \code{markdown = TRUE}}

\item{subject}{subject of email}

\item{to}{field 'to', can be character vector}

\item{cc}{field 'CC', can be character vector}

\item{bcc}{field 'BCC', can be character vector}

\item{reply_to}{field 'reply-to'}

\item{attachment}{character (vector) of file location(s), or (a \link{list} of) \link{data.frame}s. In case of file locations, these can also be SharePoint locations in the \code{\link[certeprojects:teams]{certeprojects::teams_projects_channel()}}.}

\item{header, footer}{extra text for header or footer, allows markdown if \code{markdown = TRUE}. See \code{\link[blastula:blocks]{blastula::blocks()}} to build blocks for these sections.}

\item{background}{background of the surrounding area in the email. Use \code{""}, \code{NULL} or \code{FALSE} to remove background.}

\item{send}{directly send email, \code{FALSE} will show the email in the Viewer pane and will ask whether the email should be saved to the Drafts folder of the current Microsoft 365 user.}

\item{markdown}{treat body, header and footer as markdown}

\item{signature}{text to print as email signature, or \code{NULL} to omit it, defaults to \code{\link[=get_certe_signature]{get_certe_signature()}}}

\item{automated_notice}{a \link{logical} to print a notice that the mail was sent automatically (default is \code{TRUE} if not in \code{\link[=interactive]{interactive()}} mode)}

\item{save_location}{location to save email object to, which consists of all email details and can be printed in the R console}

\item{sent_subfolder}{mail folder within Sent Items in the Microsoft 365 account, to store the mail if \code{!interactive()}}

\item{expect}{expression which should return \code{TRUE} prior to sending the email}

\item{account}{a Microsoft 365 account to use for sending the mail. This has to be an object as returned by \code{\link[=connect_outlook]{connect_outlook()}} or \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}}. Using \code{account = FALSE} is equal to setting \code{send = FALSE}.}

\item{account_teams}{a Microsoft 365 account to use for retrieving MS Teams contents. This has to be an object as returned by \code{\link[certeprojects:connect]{certeprojects::connect_teams()}} or \code{\link[Microsoft365R:client]{Microsoft365R::get_team()}}.}

\item{identifier}{a mail identifier to be printed at the bottom of the email. Defaults to \code{\link[certeprojects:project_properties]{project_identifier()}}. Use \code{FALSE} to not print an identifier.}

\item{...}{arguments for \code{\link[=mail]{mail()}}}

\item{image_path}{path of image}

\item{width}{required width of image, must be in CSS style such as "200px" or "100\%"}

\item{expr}{expression to test, an email will be sent if this expression returns an error}

\item{project_number}{Number of a project. Will be used to check the grey identifier in the email.}

\item{date}{A date, defaults to today. Will be evaluated in \code{\link[=as.Date]{as.Date()}}. Can also be of length 2 for a date range.}
}
\description{
These functions use the \code{Microsoft365R} R package to send emails via Microsoft 365. They require an Outlook Business account.
}
\details{
\code{\link[=mail_on_error]{mail_on_error()}} can be used for automated scripts.

\code{\link[=mail_plain]{mail_plain()}} sends a plain email, without markdown support and with no signature.

Use \code{\link[=mail_is_sent]{mail_is_sent()}} to check whether a project email was sent on a certain date from any Sent Items (sub)folder. The function will search for the grey identifier in the email body, which is formatted as \verb{[-yymmdd][0-9]+[-project_number][^0-9a-z]}. It returns \code{TRUE} if any email was found, and \code{FALSE} otherwise. If \code{TRUE}, the \link[=names]{name} will contain the date(s) and time(s) of the sent email.
}
\examples{
mail("test123", "test456", to = "mail@domain.com", account = NULL)

mail_plain("test123", "test456", to = "mail@domain.com", account = NULL)

mail(mail_image(image_path = system.file("test.jpg", package = "certemail")),
    "test456", to = "mail@domain.com", account = NULL)


\dontrun{

# attachments can be local files, SharePoint files, or data.frame objects
mail("test123", "test456", to = "mail@domain.com", account = NULL,
     attachment = c("C:/existing_file.pdf",
                    "Testproject - p123/sharepoint_file.pdf"))

# alternatively, use certeprojects::project_get_file() to easily add a SharePoint file:
mail("test123", "test456", to = "mail@domain.com", account = NULL,
     attachment = project_get_file("Analyse.*test.pdf", 123))
}


# data.frames as body will be transformed with certestyle::plain_html_table()
mail(body = mtcars[1:5, ],
     subject = "Check these cars!",
     to = "somebody@domain.org",
     account = FALSE)

# but better is to add as an attachment - they will become Excel files
mail(body = "Hello there",
     subject = "Check these cars!",
     to = "somebody@domain.org",
     attachment = mtcars[1:5, ],
     account = FALSE)

# use list() to add multiple data sets / attachments
mail(body = "Hello there",
     subject = "Check these cars and flowers!",
     to = "somebody@domain.org",
     attachment = list(mtcars[1:5, ],
                       iris),
     account = FALSE)
}
\seealso{
\code{\link[=download_mail_attachment]{download_mail_attachment()}}
}
