% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{mail}
\alias{mail}
\alias{mail_plain}
\alias{mail_image}
\alias{mail_on_error}
\alias{download_mail_attachment}
\title{Send emails using Microsoft 365}
\usage{
mail(
  body,
  subject,
  to,
  cc = NULL,
  bcc = read_secret("mail.auto_bcc"),
  reply_to = NULL,
  attachment = NULL,
  header = FALSE,
  footer = FALSE,
  background = colourpicker("certeblauw6"),
  send = TRUE,
  markdown = TRUE,
  signature = TRUE,
  signature_name = get_name_and_job_title(),
  signature_address = get_mail_address(),
  automated_notice = !interactive(),
  save_location = read_secret("mail.export_path"),
  expect = NULL,
  ...
)

mail_plain(
  body,
  subject,
  to,
  cc = NULL,
  bcc = read_secret("mail.auto_bcc"),
  reply_to = read_secret("mail.auto_reply_to"),
  attachment = NULL,
  send = TRUE,
  ...
)

mail_image(image_path, width = NULL, ...)

mail_on_error(expr, to = read_secret("mail.error_to"), ...)

download_mail_attachment(
  path = getwd(),
  search = NULL,
  search_subject = NULL,
  search_from = NULL,
  folder = get_inbox_name(),
  n = 50,
  sort = "received desc",
  overwrite = TRUE
)
}
\arguments{
\item{body}{body of email, allows markdown if \code{markdown = TRUE}}

\item{subject}{subject of email}

\item{to}{field 'to', can be character vector}

\item{cc}{field 'CC', can be character vector}

\item{bcc}{field 'BCC', can be character vector}

\item{reply_to}{field 'reply-to'}

\item{attachment}{character (vector) of file location(s)}

\item{header, footer}{extra text for header or footer, allows markdown if \code{markdown = TRUE}. See \code{\link[blastula:blocks]{blastula::blocks()}} to build blocks for these sections.}

\item{background}{background of the surrounding area in the email. Use \code{""}, \code{NULL} or \code{FALSE} to remove background.}

\item{send}{directly send email, \code{FALSE} will show the email in the Viewer pane and will ask whether the email should be saved to the Drafts folder of the current Microsoft 365 user.}

\item{markdown}{treat body, header and footer as markdown}

\item{signature}{add signature to email}

\item{signature_name}{signature name}

\item{signature_address}{signature email address, placed directly below \code{signature_name}}

\item{automated_notice}{print a notice that the mail was send automatically (default is \code{TRUE} is not in \code{\link[=interactive]{interactive()}} mode)}

\item{save_location}{location to save email object to, which consists of all email details and can be printed in the R console}

\item{expect}{expression which should return \code{TRUE} prior to sending the email}

\item{...}{arguments for \code{\link[=mail]{mail()}}}

\item{image_path}{path of image}

\item{width}{required width of image, must be in CSS style such as "200px" or "100\%"}

\item{expr}{expression to test, an email will be sent if this expression returns an error}

\item{path}{location to save attachment(s) in}

\item{search}{an ODATA filter, ignores \code{sort}}

\item{search_subject}{equal to \code{search = "subject:(search_subject)"}}

\item{search_from}{equal to \code{search = "from:(search_from)"}}

\item{folder}{email folder to search in, defaults to Inbox name of the current user by calling \code{\link[=get_inbox_name]{get_inbox_name()}}}

\item{n}{maximum number of emails to list}

\item{sort}{initial sorting}

\item{overwrite}{logical to indicate whether existing local files should be overwritten}
}
\description{
This uses the \code{Microsoft365R} package to send email via Microsoft 365. Connection will be made internally using \code{\link[Microsoft365R:client]{Microsoft365R::get_business_outlook()}} using the Certe tenant ID.
}
\details{
\code{\link[=mail_on_error]{mail_on_error()}} can be used for automated scripts.

\code{\link[=mail_plain]{mail_plain()}} send a plain email, without markdown support and with no signature.

If both \code{search_subject} and \code{search_from} are provided, they will be matched as 'AND'. \code{search_from} can contain any sender name or email address.
}
